refactor: enhance type safety and clarity in IBU and dataset handling
- Updated the IBU implementation to use generic type parameters for improved type safety across various functions and classes.
- Refactored the `ArrayDataset`, `Events`, `ZXY`, and `Populations` classes to support generic types, enhancing flexibility and maintainability.
- Adjusted type hints in the `Batch` type definition for consistency with the new generic approach.
- Improved documentation in the README to reflect changes in function signatures and clarify usage.

+++ src/ran/rantypes/constants.py
@@ -49,4 +49,4 @@ JET_OBS: Final[dict[str, JetVarInfo]] = {
     "sdm": JetVarInfo((-14, -2), "Soft Drop Jet Mass", r"$\ln\rho$"),
 }
 
-DEFAULT_PURITY_THRESHOLD: Final[np.double] = np.sqrt(0.5, dtype=np.double)
+DEFAULT_PURITY_THRESHOLD: Final[float] = np.sqrt(0.5)
